requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-extensions==3.2.3
django-phonenumber-field==7.3.0
phonenumbers==8.13.37
sqlparse==0.5.0
tzdata==2024.1
